refactor(all): refactor server codes
